+++ hoardcore.py
@@ -15,7 +15,7 @@ Usage:
 """
 from __future__ import annotations
 
-__version__ = "0.8.3"
+__version__ = "0.8.4"
 
 import argparse
 import asyncio